src/goLanguageServer.ts: add go.trace.server configuration
This is available by the vscode lsp client library.
We needed to set the client id - the language server client
library to look up <client_id>.trace.server. Chose 'go' as
the client id.

Acceptable values (off, messages, verbose) are from
https://github.com/microsoft/vscode-languageserver-node/blob/b1f6a443efad3d61cf83344589099dab5318ca66/jsonrpc/src/main.ts#L207

Added this setting to package.json to allow the rich setting UI to
display it properly.

Note: this tracing is done by the lsp client, so it may be
different from the server-side logging gopls provides with
-rpc.trace options.

Change-Id: I3e33a16c953c78627e779383bba56d53ae51e8e0