src/goPath: remove dependency on util that depends on vscode
golang.org/cl/239697 accidentally added the new dependency on util - to find
the value of go.goroot, which broke the debug adapter that runs outside
vscode.

Made getBinPathWithPreferredGopath accept a preferredGoroot param,
and renamed the function to reflect this change.

We need to remove any dependency from src/debugAdapter to src to avoid
a similar bug in the future.

We need to have a test, using the built extension.

Updates golang/vscode-go#137.

Change-Id: I586438f1f391aaff0c0cdc806de1e9f3fd5deba9