src/goEnvironmentStatus: remove a circular dependency
goInstallTools.ts imports goEnvironmentStatus.ts for initGoStatusBar.
goEnvironmentStatus.ts imported goInstallTools.ts for getActiveGoRoot.

The caller of getActiveGoRoot is the chooseGoEnvironment command
that is activated only after initGoStatusBar is called.
initGoStatusBar is called after updateGoVarsFromConfig, so calling
updateGoVarsFromConfig in getActiveGoRoot is unnecessary.

This CL replaces getActiveGoRoot with getCurrentGoRoot,
which removes the need to import goInstallTools.ts from
goEnvironmentStatus.ts.

Change-Id: I79320bd22fae7470c28dfd7f883a77b3a3c177a2