src/goMain.ts: move the go installation suggestion out of the critical path
This slows down the activation. One side effect is to start language server
later. This doesn't have to be on the critical path, and also updateGoVarsFromConfig
will soon do the same.

updateGoVarsFromConfig will check the go binary asynchronously
and suggest download upon detecting the issue.

And while we are here, remove the code that prevents updateGoVarsFromConfig
from running when some environment variables are set. updateGoVarsFromConfig
is called upon activation & when the relevant configuration fields are updated.
So, I suppose this is called rarely, and also it's right to reprocess and
update relevant info.

Update golang/vscode-go#363

Change-Id: I65a50915630d3d68d892ca267d248cd42958b9fb