src/goInstallTools: install -gomod tools in the right directory
Tools with the `-gomod` prefix are installed using `go build -o`
in order to avoid the conflicts with GOPATH-mode tools that have
the same name. The extension used `$GOPATH/bin/gocode-gomod`
as the -o flag value even ehen GOPATH contains multiple directories.
The go command uses whatever passed with `-o` and creates the
output directory if necessary. As a result, if the OS accepts
the output, the go command will succeed.

When GOPATH includes multiple elements, we should install
them in the first element's bin directory.

Fixes golang/vscode-go#368

Change-Id: Ief53545042056063a3bdcf57df4bf63a5a133fe2