src/goDeclaration.ts: godef requires the right `GOROOT` env var
It uses runtime.GOROOT() to clean file names. If the version of
Go used to build the godef tool is different from the version of
Go we are currently using is different, it will not know the
right place for standard libraries without GOROOT env var.

Updates golang/vscode-go#294
Updates golang/vscode-go#146

Change-Id: I4b93023634d4f20b4e9802f756e168e41c1055f0