src/debugAdapter2: launch as an external process, and fix config
Eventually we want to inline the debug adapter, but until it gets
more stable and becomes the default adapter, let's launch it
as a separate process to isolate failures. This is handled by the
definition in package.json.

In order to avoid accidental installation of the process-wide
uncaughtException handler when we switch back to the inline mode,
move the handler out of goDlvDebug.ts.

Also, fixes the default configuration provider for delve dap
debug adapter. It should use 'godlvdap' as the type.

Fixes golang/vscode-go#469
Updates golang/vscode-go#23

Change-Id: I4df2fff51c703995fd557fe5595a367d7048bd7b