src/goDebugConfiguration.ts: add resolveDebugConfiguration back
https://go-review.googlesource.com/c/vscode-go/+/248659 replaced
resolveDebugConfiguration with resolveDebugConfigurationWithSubstitutedVariables
in order to resolve the envFile correctly in case
the property requires substitution.

I thought, when debug is requested without launch.json, the default
debugConfiguration provided by provideDebugConfiguration would be
used and eventually resolveDebugConfigurationWithSubstitutedVariables
would be called to process it further. However, it doesn't seem that's the case.

Instead, I observed, resolveDebugConfiguration is called first
with an empty debugConfiguration and VS Code just drops the launch
request without calling resolveDebugConfigurationWithSubstitutedVariables
at all.

We still need to implement resolveDebugConfiguration and
let it populate the debugConfiguration.

Fixes golang/vscode-go#640

Change-Id: I693c19df1a47cb1011cccdd2f582b4cf3c0fc2ab