test/integration: fix broken updateGoVarsFromConfig tests
The test setup callback shouldn't be async. Async test setup
functions are not guaranteed to run before the tests end.

And it looks like the necessary code to mutate `go.goroot`
and `go.alternateTools` configuration for testing was removed
accidentally, but it was not detected on time because of the
use of async functions as the test setup callback functions.

Fixes golang/vscode-go#261

Change-Id: I1e77799313bdebf1373f65fab04421f296fd48dd