src/goInstallTools.ts: mutate PATH only when necessary
When the extension chooses a different version of go
than the one from the system default - because the user
has configured `go.goroot` or `go.alternateTools.go`, or
used the `Go: Choose Go Environment` command - the extension
mutates the `PATH` (or `Path` on windows) environment variable
so all the underlying tools pick the same go version.
It also changes the environment variable collection used
in the integrated terminal so when the user invokes `go`
from the terminal, they use the go version consistent
with the extension.

But this behavior can conflict with external version
management software. Change the PATH environment only
if the extension is configured to choose the go binary.

Fixes golang/vscode-go#679.
Update golang/vscode-go#544.

Change-Id: I9f7acb26b752ed33dbde2b238a67ed09616b43e5