src/debugAdapter2: supply launch in the debug session creation time
VS Code launches the thin debug adapter written in nodejs that
will eventually start or connect to  a `dlv dap` process and proxy
the interaction between the dlv dap process and the vscode.
In order to start the `dlv dap` process, this thin debug adapter
currently needs some info only available from the launch/attachRequest.
The client (vscode) will not send the launch/attachRequest until
the initializeRequest handshaking is complete. In order to answer
correctly to the initializeRequest, the thin debug adapter requires
to start or connect to the `dlv dap` process, so here we have
a chicken-and-egg problem.

This CL attempts to supply the launch/attach configuration to
the thin debug adapter during the construction time, which allows
the thin debug adapter can start/connect to the `dlv dap` process
and simply forward the initializeRequest to the `dlv dap`.

Unfortunately, this trick is only feasible when we launch the
thin debug adapter in inlined mode. vscode-debugadapter DebugSession.run
is not flexible enough for us to supply additional custom arguments
as it is. So, this CL switches to use the inline mode debug adapter.

Alternative approaches to be discussed are

  - Keep the thin debug adapter respond with a minimal set of
    capability, and let the `dlv dap` send Capabilities events to
    the client  with the updated capability info later. The spec
    states that the Capabilities events should occur early enough
    and there is no guarantee on whether the clients will honor
    the updated capabilities. So, it is still unclear whether this
    will work for all the capabilities and when is the good time
    for the dlv dap to send these events.

    (note) It seems like we are not forwarding the client's
    initializeRequest to `dlv dap` after starting the process.
    If that's true, it is a bug - the DAP server needs to learn
    about the client's capabilities too.

  - Change `dlv dap` not to require any launch/attach configuration
    dependent information to start. Currently the thin adapter uses
    host/port, showLog, logOutput, cwd properties to invoke `dlv dap`.
    (We can omit cwd, and the purpose of host/port in the local&launch
    mode is unclear)

  - Look into a way to extend DebugSession.run to accept custom
    args and pass the args/flags to use when invoking `dlv dap`.
    This will allow us to launch the thin debug adapter as a separate
    process.

    (note) In case of local launch, we can start the `dlv dap` process
    from the DA descriptor factory before launching the thin debug
    adapter and have the thin adapter just connect to it. But this
    makes the process management a bit more involved - if the debug
    session ends we need to make sure to clean up the dlv dap process.

  - Should we revisit our decision on the thin adapter? The original
    motivation was to allow vscode-specific functionalities such as
    automatic path mapping. But given that a large part of the automatic
    path mapping is currently depending on the delve's JSON-RPC API
    (ListPackageBuildInfo, ListSources, ...) I am getting skeptical.
    Other than the path mapping, what other vscode-specific features
    would find this thin adapter useful?

DO NOT SUBMIT

For discussion about golang/vscode-go#822

Change-Id: I82fa3add57495a79d05d2de902a01244b9effa71