src/testUtils: fix test streaming output handling
The intention of https://go-review.googlesource.com/c/vscode-go/+/246918
was to enable `-json` only if the user already uses `-v` because `-json`
automatically enables the verbose mode and `-v` is the way to activate
the streaming test output feature in the go command.

By mistake we scanned the argument list before the user specified flag
list was merged. As a result, `-json` flag was never added.  This change
corrects the bug.

This time, we add tests to avoid regression. In order to make testing
easier this CL restructures the goTest function: 1) moves test target
argument computation to the getTestTargetPackages, and 2) moves the test
command generation into the computeTestCommand function.

We also fix a bug that caused the flags located after `-args` to influence
our command line computation (-run, -tags). Everything after `-args`
should be considered user's custom arg and  passed as it is without
modification.

Fixes #900
Updates #316

Change-Id: Idc11a1ba5d3f5f4cf5467192801c539123b8c241