.vscode/launch.json: unset test guard env var
Bug (microsoft/vscode#83187) is causing VSCODE_GO_IN_TEST env var
to leak and subsequent extension tests to fail to activate fully.
This is a short-term workaround until we find a better solution
to deactivate the extension during testing or learn how to test
a bundled extension better (microsoft/vscode-extension-samples#363).

Change-Id: If5b9b56ef8360547c518240a42286c9385b53b44