.github/workflows: install stable delve version for testing
The extension maintains two dlv binaries for debugging, a stable
version of dlv for the legacy adapater and dlv built from master
to be used with dlv dap, which is under active development. For
testing we should use the stable version of delve for the legacy,
and only use master for dlv-dap.

Updates golang/vscode-go#1456

Change-Id: Ie4482aebb4fee1579896702faca207182b4746c3