src/language/goLanguageServer: do not require gopls when language server is disabled
This fixes a bug that made the extension to look up the path to the gopls
binary when go.useLanguageServer is false. During the lookup, if the gopls wasn't
available, getLanguageServerToolPath eventually triggerred the missing tool prompt
and that was not appropriate. Return from buildLanguageServerConfig early
if we don't need the language server.

Updates golang/vscode-go#2300

Change-Id: I7572244350f953c8ce874086f4f5cbb45a7301af