test/integration/goDebug: skip redundant tests in withConsole mode
integrated terminal mode testing doesn't need to run the full test
suite. It is sufficient to check message forwarding through the proxy
thin adapter is working. So, skip most of irrelevant tests that
are tested in the default mode.

And avoid printing the 'env' property that floods the test output
and may leak too much information in logging.

Change-Id: I83ce54413cc9ca291e4b5237fca47bc3a2ac92b0