test/integration/goDebug: speed up setUpRemoteProgram
Instead of sleeping for fixed 10sec, watch the stdout
and look for the server start signature message.

Either Delve's start message (if --continue flag is not used)
   i.e., DAP server listening at: or API server listening at:
or the helloWorldServer's log message (if --continue flag is used)
   helloWorldServer starting to listen on

The logic is inspired from the current debug adapter's
dlv spawning logic.

In most cases, this approach makes tests complete fast.

If server takes more than 10sec to start up, this
approach is more robust as long as the server setup can be done
within hard 30sec limit.

Change-Id: Ie880bdf50cb33e75873e2d8fb4e92ca9c53e633f