test/gopls: narrow the scope of fake formatTool setting
The extension supports custom formatter. When gopls is used, this
is done by intercepting the formatting request from the middleware
and passing it to the legacy format provider that calls the specified
tool. The "Nonexistent formatter" test is meant to take this
middleware code path and let the legacy format provider fail by
setting `go.formatTool` with a bogus, non-existing tool name.

This setting doesn't need to affect other tests though.
Limit this setting to only the nonexistent formatter, by starting
a new language server instance for each test.

Additionally this fixes bugs:
- wait to show the document to format - the legacy formatter doesn't
  run if there is no visible document. (I don't know why we don't format
  non-visible document but that's how it's written goFormat.ts:24).
- close the open editor after each test.

For golang/vscode-go#1603

Change-Id: I3486b299aa2a2660fb971fed54d3f941be13698d