test/unit/goDebug: delete the goDebug test
This test imports the legacy DA module in src/debugAdapter/goDebug.
Even though the tested functions look simple, loading this module
has side effect - running the debug adapter process - and this
uncleaned job makes the unit test hang.

This code remained stable for a while but was running only by the
legacy DA. We are about to delete the legacy DA. So I don't see
much value of keeping this test any longer. Remove this test.

Change-Id: I97706c7a60e101b51ff9356c7a0e958d4b44429c