build: run tests and build golang.go-nightly
Building and testing the extension requires a custom docker
image that includes Go, Node.js, jq, gopls, dlv, staticcheck, etc.

build-ci-image.yaml is a workflow that builds and stores the image
in Artifact Registry. (vscode-go-docker-repo/ci-image).

release-nightly.yaml workflow uses the container.

- The workflow clones the vscode-go repo (master branch).
- Before packaging, the "prepare nightly release" step
  adjusts the package.json for golang.go-nightly use.
- Then run tests. It's possible that the lint test and the
  tools/generate.go test may be unhappy about the spacing and formatting
  caused by the modification made during the previous step. However,
  the errors are not critical for packaging/publishing. Refactoring of
  all.bash was to skip the lint, generate.go tests during nightly
  release.

The workflow uploads the built extension to the project's GCS bucket but
does not publish the built extension to the marketplace yet.

Change-Id: Ib48fdc45f1fd1c86fcfc2d293e7f60690b2cad11