src/goInstallTools: require go1.16 or newer for tools installation
This deletes code that handled tools installation using `go get` and
GOPATH mode. The recent survey indicates <<1% of users use go1.15 or
older.

And, update README.md Getting started section to suggest go1.18 or
newer. When go1.21 is released, we can tighten it to go1.19 or newer.

Fixes golang/vscode-go#2898

Change-Id: Ic7ed2218cf6c5a0db0e8851a388ab184f8bd8654