src/goMain: apply substituteEnv when reading go.goroot
I think a better place to do this is when resolvePath is called
or even when getGoConfig returns the goroot values.
But resolvePath is used in many other places so I want to proceed
more conservatively.
Applying in getGoConfig is tricky since it is currently returning
a readonly vscode workspace configuration object, and the caller
accesses the value in multiple, different ways.

Fixes golang/vscode-go#3006

Change-Id: I3733c0ecdcabd3d4828f666f9a5b87aa877971f9