vscgo: add a little helper program for vscode-go
github.com/golang/vscode-go/vscgo is a small go program
that the extension uses to delegate some tasks.
The program will be installed in the extension path
and its behavior will be tightly tied with the vscode
go extension version.

From v0.41.x the extension will use this to record go
telemetry counters defined in the extension. The extension
spawns a vsccode inc_counters process periodically and
passes accumulated counters (key/value pairs) to the program.

We plan to expand the functionality in the future,
such as tools installation, upgrade, developer
survey config check, that are not latency critical
and easier to write in Go.

For golang/vscode-go#3023
For golang/vscode-go#2891

Change-Id: I91427fc6a06dc2781ab21c12e0b8ed5957cba9b1