extension/test/integration: disable local dlv install tests
For go1.19 and go1.20, the extension tries to install a pinned version
of dlv (the last version that is compatible with the old go version).
The local proxy used for testing does not have the pinned version
so the test fails. Ironically this test failure indicates the dlv install
logic works as expected. Since the fix is non-trivial and hard to be
done before v0.42.0 release, skip them in go1.19/go1.20 CI.

FORCE RUN CI

Change-Id: Iccd9d905ab46adf255c52d82817046d4fb85fe81