extension/src/goInstallTools.ts: require go1.21+ for tools installation
The Go extension will require go1.21 for tools installation from v0.44.0
(and is prerelease version v0.43.x).

This is a planned change and it was discussed in the v0.42.0 release note.
(https://github.com/golang/vscode-go/releases/tag/v0.42.0 Jul 17 2024).

`installTools` is the entry function for tools installation.
If the go version is too old, it suggests go1.21+ or the workaround
(go.toolsManagement.go).

* Misc changes
 - Previously, when the build info of a binary is not available,
   we didn't ask to update the tool. Since go1.18, the build info
   should be available. So, now suggest to reinstall the tool.
 - Bug fix: For vscgo, we used toolExecutionEnvironment when running
   go install.
   It should be toolInstallationEnvironment. This clears some env vars
   like GO111MODULE, GOPROXY, GOOS, GOARCH, GOROOT which can interfere
   with the go tool invocation.

Fixes golang/vscode-go#3411

Change-Id: Ifff0661d88a9adfc6bd3e0a25702d91921bcb77f