extension/src/goTools.ts: remove 'replacedByGopls' field
FORCE RUN CI

This field was added when we were deprecating several tools with gopls
gradually and we wanted to keep the complexity of `getConfiguredTools`
manageable. Now we have only a handful number of tools. Remove it.

Also, remove the custom formatter check in getConfiguredTools.
The idea was to add a custom formatter tool to the configured tools list
if the user configured "go.formatTool": "custom". But, since custom tool
name won't be in `allToolsInformation`, so the custom tool won't be
included in the list anyway. (We install only tools known to us).

Change-Id: I4b26d3756fe45e1cc29f8434b1b5d83ee8977d47